local_regression/requirements.txt
absl-py==0.8.0
alabaster==0.7.10
asn1crypto==0.24.0
astor==0.8.0
astroid==1.6.1
astropy==2.0.3
attrs==17.4.0
Babel==2.5.3
certifi==2016.2.28
cffi==1.12.3
chardet==3.0.4
chumpy==0.69
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
datashape==0.5.4
decorator==4.2.1
distributed==1.20.2
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.13
gast==0.3.2
gevent==1.2.2
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.12
grpcio==1.12.1
h5py==2.10.0
heapdict==1.0.0
html5lib==0.9999999
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
matplotlib==2.1.2
mccabe==0.6.1
mistune==0.8.3
mpmath==1.0.0
msgpack-python==0.5.1
natsort==7.1.1
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.36.2
numexpr==2.6.4
numpy==1.14.2
numpydoc==0.7.0
odo==0.5.1
olefile==0.46
opencv-python==4.1.2.30
openpyxl==2.4.10
packaging==16.8
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.4.1
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.4.0
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.2
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.3.2
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
rope==0.10.7
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.12.0
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==4.5.3
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
zict==0.1.3
